+++ src/main/java/sesac/server/feed/service/NoticeService.java
@@ -1,0 +1,14 @@
+package sesac.server.feed.service;
+
+import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.log4j.Log4j2;
+import org.springframework.stereotype.Service;
+
+@Log4j2
+@Service
+@Transactional
+@RequiredArgsConstructor
+public class NoticeService {
+
+}
