+++ src/main/java/sesac/server/feed/service/PostService.java
@@ -20,14 +20,10 @@ import sesac.server.feed.dto.response.ReplyResponse;
 import sesac.server.feed.entity.ArticleType;
 import sesac.server.feed.entity.FeedType;
 import sesac.server.feed.entity.Hashtag;
-import sesac.server.feed.entity.Likes;
-import sesac.server.feed.entity.Notice;
 import sesac.server.feed.entity.Post;
 import sesac.server.feed.entity.PostHashtag;
 import sesac.server.feed.exception.PostErrorCode;
 import sesac.server.feed.repository.HashtagRepository;
-import sesac.server.feed.repository.LikesRepository;
-import sesac.server.feed.repository.NoticeRepository;
 import sesac.server.feed.repository.PostHashtagRepository;
 import sesac.server.feed.repository.PostRepository;
 import sesac.server.user.entity.User;
@@ -41,11 +37,9 @@ import sesac.server.user.repository.UserRepository;
 public class PostService {
 
     private final PostRepository postRepository;
-    private final NoticeRepository noticeRepository;
     private final UserRepository userRepository;
     private final HashtagRepository hashtagRepository;
     private final PostHashtagRepository postHashtagRepository;
-    private final LikesRepository likesRepository;
 
     public Post createPost(Long userId, String feedType, CreatePostRequest request) {
         User user = userRepository.findById(userId)
@@ -134,84 +128,9 @@ public class PostService {
         postRepository.delete(post);
     }
 
-    @Transactional
-    public void likeFeed(CustomPrincipal principal, Long feedId, ArticleType articleType) {
-        boolean alreadyLiked = isLiked(principal.id(), feedId, articleType);
-        if (alreadyLiked) {
-            throw new BaseException(PostErrorCode.ALREADY_LIKED);
-        }
-
-        Object feed = getFeedById(feedId, articleType);
-        User user = userRepository.getReferenceById(principal.id());
-
-        Likes like = getLikes(feed, user, articleType);
-
-        likesRepository.save(like);
-    }
-
-    @Transactional
-    public void cancelLikeFeed(CustomPrincipal principal, Long feedId, ArticleType articleType) {
-        boolean isLiked = isLiked(principal.id(), feedId, articleType);
-        if (!isLiked) {
-            throw new BaseException(PostErrorCode.NOT_LIKED);
-        }
-
-        Object feed = getFeedById(feedId, articleType);
-        User user = userRepository.getReferenceById(principal.id());
-
-        deleteLike(feed, user, articleType);
-    }
-
     private boolean hasPermission(CustomPrincipal principal, Long userId) {
         return principal.role().equals(UserRole.MANAGER.toString()) ||
                 principal.id().equals(userId);
-    }
-
-    private boolean isLiked(Long userId, Long feedId, ArticleType articleType) {
-        return switch (articleType) {
-            case POST ->
-                    likesRepository.existsByUserIdAndPostIdAndType(userId, feedId, articleType);
-            case NOTICE ->
-                    likesRepository.existsByUserIdAndNoticeIdAndType(userId, feedId, articleType);
-            default -> throw new IllegalArgumentException("없는 피드 타입입니다.");
-        };
-    }
-
-    private Object getFeedById(Long feedId, ArticleType articleType) {
-        switch (articleType) {
-            case POST:
-                return postRepository.findById(feedId)
-                        .orElseThrow(() -> new BaseException(PostErrorCode.NO_POST));
-            case NOTICE:
-                return noticeRepository.findById(feedId)
-                        .orElseThrow(() -> new BaseException(PostErrorCode.NO_NOTICE));
-            default:
-                throw new IllegalArgumentException("없는 피드 타입입니다.");
-        }
-    }
-
-    private Likes getLikes(Object feed, User user, ArticleType articleType) {
-        Likes.LikesBuilder builder = Likes.builder()
-                .user(user)
-                .type(articleType);
-
-        if (feed instanceof Post) {
-            return builder.post((Post) feed).build();
-        } else if (feed instanceof Notice) {
-            return builder.notice((Notice) feed).build();
-        } else {
-            throw new IllegalArgumentException("없는 피드 타입입니다.");
-        }
-    }
-
-    private void deleteLike(Object feed, User user, ArticleType articleType) {
-        if (feed instanceof Post) {
-            likesRepository.deleteByUserAndPostAndType(user, (Post) feed, articleType);
-        } else if (feed instanceof Notice) {
-            likesRepository.deleteByUserAndNoticeAndType(user, (Notice) feed, articleType);
-        } else {
-            throw new IllegalArgumentException("없는 피드 타입입니다.");
-        }
     }
 
     private FeedType getEnumType(String value) {
